adding docs directory, and setting up modules
