+++ docs/index.rst
@@ -1,0 +1,22 @@
+.. HPCS documentation master file, created by
+   sphinx-quickstart on Tue Mar 19 13:07:16 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to HPCS's documentation!
+================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+   
+   modules
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
